@@ -1,6 +1,5 @@
 plugins {
-    id("com.github.johnrengelman.shadow") version "7.1.2"
-    id("io.micronaut.application") version "3.7.10"
+    id("io.micronaut.application") version "4.6.1"
 }
 
 version = "0.1"
@@ -12,16 +11,15 @@ repositories {
 
 dependencies {
     annotationProcessor("info.picocli:picocli-codegen")
-    implementation("info.picocli:picocli")
-    implementation("io.micronaut:micronaut-jackson-databind")
+    implementation("io.micronaut.serde:micronaut-serde-jackson")
     implementation("io.micronaut.picocli:micronaut-picocli")
     implementation("jakarta.annotation:jakarta.annotation-api")
     runtimeOnly("ch.qos.logback:logback-classic")
-    implementation("io.micronaut:micronaut-validation")
+    implementation("io.micronaut.validation:micronaut-validation")
 
-    testImplementation("io.micronaut:micronaut-http-client")
+    testImplementation("io.micronaut:micronaut-http-client-jdk")
 
-    implementation("org.apache.logging.log4j:log4j-core:2.15.0")
+    implementation("org.apache.logging.log4j:log4j-core:2.24.0")
 }
 
 
@@ -29,8 +27,9 @@ application {
     mainClass.set("com.example.Demo2Command")
 }
 java {
-    sourceCompatibility = JavaVersion.toVersion("11")
-    targetCompatibility = JavaVersion.toVersion("11")
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17)
+    }
 }
 
 micronaut {
