@@ -18,8 +18,7 @@
  */
 
 plugins {
-    id("com.github.johnrengelman.shadow") version "7.1.2"
-    id("io.micronaut.application") version "3.5.1"
+    id("io.micronaut.application") version "4.6.1"
 }
 
 version = "0.1"
@@ -31,27 +30,30 @@ repositories {
 
 dependencies {
     annotationProcessor("io.micronaut:micronaut-http-validation")
-    implementation("io.micronaut:micronaut-http-client")
-    implementation("io.micronaut:micronaut-jackson-databind")
+    implementation("io.micronaut:micronaut-http-client-jdk")
+    implementation("io.micronaut.serde:micronaut-serde-jackson")
     implementation("jakarta.annotation:jakarta.annotation-api")
     runtimeOnly("ch.qos.logback:logback-classic")
-    implementation("io.micronaut:micronaut-validation")
 
-    implementation("org.apache.logging.log4j:log4j-core:2.17.0")
+    implementation("io.micronaut.validation:micronaut-validation")
+    implementation("org.apache.logging.log4j:log4j-core:2.24.3")
 }
 
 
 application {
     mainClass.set("com.example.Application")
 }
+
 java {
-    sourceCompatibility = JavaVersion.toVersion("11")
-    targetCompatibility = JavaVersion.toVersion("11")
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17)
+    }
 }
 
 graalvmNative.toolchainDetection = false
 micronaut {
-    runtime("netty")
+    version "4.9.4"
+    runtime("http_server_jdk")
     testRuntime("junit5")
     processing {
         incremental(true)
