@@ -18,8 +18,7 @@
  */
 
 plugins {
-    id("com.github.johnrengelman.shadow") version "7.1.2"
-    id("io.micronaut.application") version "3.5.1"
+    id("io.micronaut.application") version "4.6.1"
 }
 
 version = "0.1"
@@ -32,14 +31,17 @@ repositories {
 application {
     mainClass.set("com.example.Application")
 }
+
 java {
-    sourceCompatibility = JavaVersion.toVersion("11")
-    targetCompatibility = JavaVersion.toVersion("11")
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17)
+    }
 }
 
 graalvmNative.toolchainDetection = false
 micronaut {
-    runtime("netty")
+    version "4.9.4"
+    runtime("http_server_jdk")
     testRuntime("junit5")
     processing {
         incremental(true)
